Implement fully functioning bug delete functions

+++ cyberbug_frontend/src/components/project/bugDetails.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { hideToggle, showToggle } from "../../actions/toggleActions";
 import { setActiveBug, editBug, deleteBug } from "../../actions/bugActions";
+import { Redirect } from "react-router";
 
 class BugDetails extends Component {
   state = {
@@ -9,35 +10,46 @@ class BugDetails extends Component {
   };
 
   setBug() {
-    return this.props.bugs.filter((bug) => {
-      return bug.bugId === this.props.match.params.bugId;
-    })[0];
+    if (this.props.bugs) {
+      return this.props.bugs.filter((bug) => {
+        return bug.bugId === this.props.match.params.bugId;
+      })[0];
+    } else {
+      return null;
+    }
   }
 
   renderBugDetails() {
+    console.log(this.props.match.url.slice(0, -7));
+
     if (this.props.bugs) {
       const bug = this.setBug();
-      return (
-        <div>
-          {/* TODO: This will have all of the detailed info */}
-          <div className="centered-flex" id="details-title">
-            <h3>{bug.bugTitle}</h3>
+      // console.log(bug);
+      if (bug) {
+        return (
+          <div>
+            {/* TODO: This will have all of the detailed info */}
+            <div className="centered-flex" id="details-title">
+              <h3>{bug.bugTitle}</h3>
+            </div>
+            <p>Status: {bug.bugStatus}</p>
+            <p>Description:</p>
+            <div className="test-border">{bug.bugDescription}</div>
+            <p>Details:</p>
+            <div className="test-border">{bug.bugDetails}</div>
+            <p>Tags:</p>
+            <div className="test-border">{bug.bugTags}</div>
+            <p>Due Date: {bug.bugDueDate}</p>
+            <p>Line No: {bug.bugLineNo}</p>
+            <p>Priority: {bug.bugPriority}</p>
+            <p>Created: {bug.created_at}</p>
+            <p>Updated: {bug.updated_at}</p>
+            <button onClick={this.handleOnClickDelete}>Delete</button>
           </div>
-          <p>Status: {bug.bugStatus}</p>
-          <p>Description:</p>
-          <div className="test-border">{bug.bugDescription}</div>
-          <p>Details:</p>
-          <div className="test-border">{bug.bugDetails}</div>
-          <p>Tags:</p>
-          <div className="test-border">{bug.bugTags}</div>
-          <p>Due Date: {bug.bugDueDate}</p>
-          <p>Line No: {bug.bugLineNo}</p>
-          <p>Priority: {bug.bugPriority}</p>
-          <p>Created: {bug.created_at}</p>
-          <p>Updated: {bug.updated_at}</p>
-          <button onClick={this.handleOnClickDelete}>Delete</button>
-        </div>
-      );
+        );
+      } else {
+        return <Redirect to={this.props.match.url.slice(0, -7)} />;
+      }
     } else {
       <h3>Loading Info...</h3>;
     }
@@ -80,20 +92,22 @@ class BugDetails extends Component {
   };
 
   handleOnClickSquash = (event) => {
-    console.log("Squash button clicked");
+    // console.log("Squash button clicked");
     const bug = this.props.activeBug;
     bug.bugStatus = "Complete";
     bug.bugComplete = true;
 
-    console.log(bug);
+    // console.log(bug);
     this.props.editBug({ bug: bug });
     // dispatch action to set status to complete
     // this.props.completeBug(this.props.activeBug);
   };
 
   handleOnClickDelete = (event) => {
-    console.log("delete button clicked");
-    this.props.deleteBug(this.props.activeBug);
+    // console.log("delete button clicked");
+    if (window.confirm("Are you sure you want to yeet this bug?")) {
+      this.props.deleteBug(this.props.activeBug);
+    }
   };
 }
